all tests passing

+++ src/Controller/CategoryController.php
@@ -12,55 +12,43 @@ namespace App\Controller;
 
 use App\Entity\Category;
 use App\Form\CategoryType;
+use App\Repository\CategoryRepository;
+
 use App\Repository\ClippingRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Knp\Bundle\PaginatorBundle\Definition\PaginatorAwareInterface;
 use Nines\UtilBundle\Controller\PaginatorTrait;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * Category controller.
- *
  * @Route("/category")
  */
 class CategoryController extends AbstractController implements PaginatorAwareInterface {
     use PaginatorTrait;
 
     /**
-     * Lists all Category entities.
-     *
      * @Route("/", name="category_index", methods={"GET"})
-     *
-     * @Template
      */
-    public function indexAction(Request $request, EntityManagerInterface $em) : array {
-        $qb = $em->createQueryBuilder();
-        $qb->select('e')->from(Category::class, 'e')->orderBy('e.id', 'ASC');
-        $query = $qb->getQuery();
+    public function index(Request $request, CategoryRepository $categoryRepository) : Response {
+        $query = $categoryRepository->indexQuery();
+        $pageSize = (int) $this->getParameter('page_size');
+        $page = $request->query->getint('page', 1);
 
-        $categories = $this->paginator->paginate($query, $request->query->getint('page', 1), 24);
-
-        return [
-            'categories' => $categories,
-        ];
+        return $this->render('category/index.html.twig', [
+            'categories' => $this->paginator->paginate($query, $page, $pageSize),
+        ]);
     }
 
     /**
-     * Creates a new Category entity.
-     *
      * @Route("/new", name="category_new", methods={"GET", "POST"})
      * @IsGranted("ROLE_CONTENT_ADMIN")
-     * @Template
-     *
-     * @return array|RedirectResponse
      */
-    public function newAction(Request $request, EntityManagerInterface $em) {
+    public function new(Request $request, EntityManagerInterface $em) : Response {
         $category = new Category();
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
@@ -68,71 +56,64 @@ class CategoryController extends AbstractController implements PaginatorAwareInt
         if ($form->isSubmitted() && $form->isValid()) {
             $em->persist($category);
             $em->flush();
-
-            $this->addFlash('success', 'The new category was created.');
+            $this->addFlash('success', 'The new category has been saved.');
 
             return $this->redirectToRoute('category_show', ['id' => $category->getId()]);
         }
 
-        return [
+        return $this->render('category/new.html.twig', [
             'category' => $category,
             'form' => $form->createView(),
-        ];
+        ]);
     }
 
     /**
-     * Finds and displays a Category entity.
-     *
      * @Route("/{id}", name="category_show", methods={"GET"})
-     *
-     * @Template
      */
-    public function showAction(Request $request, Category $category, ClippingRepository $repo) : array {
+    public function show(Request $request, Category $category, ClippingRepository $repo) : Response {
+        $pageSize = (int) $this->getParameter('page_size');
         $query = $repo->categoryQuery($category);
-        $clippings = $this->paginator->paginate($query, $request->query->getint('page', 1), 24);
+        $clippings = $this->paginator->paginate($query, $request->query->getint('page', 1), $pageSize);
 
-        return [
+        return $this->render('category/show.html.twig', [
             'category' => $category,
             'clippings' => $clippings,
-        ];
+        ]);
     }
 
     /**
-     * Displays a form to edit an existing Category entity.
-     *
-     * @Route("/{id}/edit", name="category_edit", methods={"GET", "POST"})
      * @IsGranted("ROLE_CONTENT_ADMIN")
-     * @Template
-     *
-     * @return array|RedirectResponse
+     * @Route("/{id}/edit", name="category_edit", methods={"GET", "POST"})
      */
-    public function editAction(Request $request, Category $category, EntityManagerInterface $em) {
-        $editForm = $this->createForm(CategoryType::class, $category);
-        $editForm->handleRequest($request);
+    public function edit(Request $request, Category $category, EntityManagerInterface $em) : Response {
+        $form = $this->createForm(CategoryType::class, $category);
+        $form->handleRequest($request);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
+        if ($form->isSubmitted() && $form->isValid()) {
             $em->flush();
-            $this->addFlash('success', 'The category has been updated.');
+            $this->addFlash('success', 'The updated category has been saved.');
 
             return $this->redirectToRoute('category_show', ['id' => $category->getId()]);
         }
 
-        return [
+        return $this->render('category/edit.html.twig', [
             'category' => $category,
-            'edit_form' => $editForm->createView(),
-        ];
+            'edit_form' => $form->createView(),
+        ]);
     }
 
     /**
-     * Deletes a Category entity.
-     *
-     * @Route("/{id}/delete", name="category_delete", methods={"GET"})
      * @IsGranted("ROLE_CONTENT_ADMIN")
+     * @Route("/{id}", name="category_delete", methods={"DELETE"})
      */
-    public function deleteAction(Request $request, Category $category, EntityManagerInterface $em) : RedirectResponse {
-        $em->remove($category);
-        $em->flush();
-        $this->addFlash('success', 'The category was deleted.');
+    public function delete(Request $request, Category $category, EntityManagerInterface $em) : RedirectResponse {
+        if ($this->isCsrfTokenValid('delete' . $category->getId(), $request->request->get('_token'))) {
+            $em->remove($category);
+            $em->flush();
+            $this->addFlash('success', 'The category has been deleted.');
+        } else {
+            $this->addFlash('warning', 'The security token was not valid.');
+        }
 
         return $this->redirectToRoute('category_index');
     }
