+++ src/Controller/SourceController.php
@@ -10,57 +10,45 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
-use App\Entity\Clipping;
 use App\Entity\Source;
 use App\Form\SourceType;
+use App\Repository\ClippingRepository;
+use App\Repository\SourceRepository;
+
 use Doctrine\ORM\EntityManagerInterface;
 use Knp\Bundle\PaginatorBundle\Definition\PaginatorAwareInterface;
 use Nines\UtilBundle\Controller\PaginatorTrait;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * Source controller.
- *
  * @Route("/source")
  */
 class SourceController extends AbstractController implements PaginatorAwareInterface {
     use PaginatorTrait;
 
     /**
-     * Lists all Source entities.
-     *
      * @Route("/", name="source_index", methods={"GET"})
-     *
-     * @Template
      */
-    public function indexAction(Request $request, EntityManagerInterface $em) : array {
-        $qb = $em->createQueryBuilder();
-        $qb->select('e')->from(Source::class, 'e')->orderBy('e.id', 'ASC');
-        $query = $qb->getQuery();
+    public function index(Request $request, SourceRepository $sourceRepository) : Response {
+        $query = $sourceRepository->indexQuery();
+        $pageSize = (int) $this->getParameter('page_size');
+        $page = $request->query->getint('page', 1);
 
-        $sources = $this->paginator->paginate($query, $request->query->getint('page', 1), 24);
-
-        return [
-            'sources' => $sources,
-        ];
+        return $this->render('source/index.html.twig', [
+            'sources' => $this->paginator->paginate($query, $page, $pageSize),
+        ]);
     }
 
     /**
-     * Creates a new Source entity.
-     *
      * @Route("/new", name="source_new", methods={"GET", "POST"})
      * @IsGranted("ROLE_CONTENT_ADMIN")
-     *
-     * @Template
-     *
-     * @return array|RedirectResponse
      */
-    public function newAction(Request $request, EntityManagerInterface $em) {
+    public function new(Request $request, EntityManagerInterface $em) : Response {
         $source = new Source();
         $form = $this->createForm(SourceType::class, $source);
         $form->handleRequest($request);
@@ -68,75 +56,64 @@ class SourceController extends AbstractController implements PaginatorAwareInter
         if ($form->isSubmitted() && $form->isValid()) {
             $em->persist($source);
             $em->flush();
-
-            $this->addFlash('success', 'The new source was created.');
+            $this->addFlash('success', 'The new source has been saved.');
 
             return $this->redirectToRoute('source_show', ['id' => $source->getId()]);
         }
 
-        return [
+        return $this->render('source/new.html.twig', [
             'source' => $source,
             'form' => $form->createView(),
-        ];
+        ]);
     }
 
     /**
-     * Finds and displays a Source entity.
-     *
      * @Route("/{id}", name="source_show", methods={"GET"})
-     *
-     * @Template
      */
-    public function showAction(Request $request, Source $source, EntityManagerInterface $em) : array {
-        $repo = $em->getRepository(Clipping::class);
+    public function show(Request $request, Source $source, ClippingRepository $repo) : Response {
+        $pageSize = (int) $this->getParameter('page_size');
         $query = $repo->sourceQuery($source);
+        $clippings = $this->paginator->paginate($query, $request->query->getint('page', 1), $pageSize);
 
-        $clippings = $this->paginator->paginate($query, $request->query->getint('page', 1), 24);
-
-        return [
+        return $this->render('source/show.html.twig', [
             'source' => $source,
             'clippings' => $clippings,
-        ];
+        ]);
     }
 
     /**
-     * Displays a form to edit an existing Source entity.
-     *
-     * @Route("/{id}/edit", name="source_edit", methods={"GET", "POST"})
      * @IsGranted("ROLE_CONTENT_ADMIN")
-     *
-     * @Template
-     *
-     * @return array|RedirectResponse
+     * @Route("/{id}/edit", name="source_edit", methods={"GET", "POST"})
      */
-    public function editAction(Request $request, Source $source, EntityManagerInterface $em) {
-        $editForm = $this->createForm(SourceType::class, $source);
-        $editForm->handleRequest($request);
+    public function edit(Request $request, Source $source, EntityManagerInterface $em) : Response {
+        $form = $this->createForm(SourceType::class, $source);
+        $form->handleRequest($request);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
+        if ($form->isSubmitted() && $form->isValid()) {
             $em->flush();
-            $this->addFlash('success', 'The source has been updated.');
+            $this->addFlash('success', 'The updated source has been saved.');
 
             return $this->redirectToRoute('source_show', ['id' => $source->getId()]);
         }
 
-        return [
+        return $this->render('source/edit.html.twig', [
             'source' => $source,
-            'edit_form' => $editForm->createView(),
-        ];
+            'edit_form' => $form->createView(),
+        ]);
     }
 
     /**
-     * Deletes a Source entity.
-     *
-     * @Route("/{id}/delete", name="source_delete", methods={"GET"})
      * @IsGranted("ROLE_CONTENT_ADMIN")
+     * @Route("/{id}", name="source_delete", methods={"DELETE"})
      */
-    public function deleteAction(Request $request, Source $source, EntityManagerInterface $em) : RedirectResponse {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($source);
-        $em->flush();
-        $this->addFlash('success', 'The source was deleted.');
+    public function delete(Request $request, Source $source, EntityManagerInterface $em) : RedirectResponse {
+        if ($this->isCsrfTokenValid('delete' . $source->getId(), $request->request->get('_token'))) {
+            $em->remove($source);
+            $em->flush();
+            $this->addFlash('success', 'The source has been deleted.');
+        } else {
+            $this->addFlash('warning', 'The security token was not valid.');
+        }
 
         return $this->redirectToRoute('source_index');
     }
